+++ src/components/CSVUploadComponent/CSVUploadComponent.tsx
@@ -1,15 +1,9 @@
-import { CSVUploadState } from "@customTypes/enums";
 import CSVUploadDatagrid from "@components/datagrids/CSVUploadDatagrid/CSVUploadDatagrid";
 import Papa from "papaparse";
 import { Column } from "@components/datagrids/GenericResponsiveDatagrid/FullScreenTable";
-import dayjs from "dayjs";
 import CSVUploadProvider from "@components/CSVUploadProvider/CSVUploadProvider";
 import { useState } from "react";
-
-interface CSVData<T> {
-  data: T;
-  state: CSVUploadState;
-}
+import { CSVData } from "./CSVHelper";
 
 interface CSVUploadComponentProps<T> {
   parseCSVData: (data: Papa.ParseResult<unknown>) => T[];
@@ -19,10 +13,6 @@ interface CSVUploadComponentProps<T> {
   validateDataRow: (data: T) => Promise<boolean>;
   key: string;
 }
-
-const convertDateFormat = (dateStr: string) => {
-  return dayjs(dateStr, "DD.MM.YYYY").format("YYYY-MM-DD");
-};
 
 const CSVUploadComponent = <T extends Record<string, unknown>>({
   parseCSVData,
@@ -61,5 +51,3 @@ const CSVUploadComponent = <T extends Record<string, unknown>>({
 };
 
 export default CSVUploadComponent;
-export type { CSVData };
-export { convertDateFormat };
